Employee-management-System-with-Angular-and-spring-boot-main/angular-frontend/src/app/employee.service.ts
import { Injectable } from '@angular/core';
import { HttpClient } from '@angular/common/http'
import { Observable } from 'rxjs';
import { Employee } from './employee';

@Injectable({
  providedIn: 'root'
})
export class EmployeeService {

  private baseURL = "http://localhost:8080/api/v1/employees";
  employee: Array<object> = [                                       // Add employee object
    {
      id: '1',
      firstName: 'Jane',
      lastName: 'Roy',
      emailId: 'janeroy@gmail.com'
    },
    {
      id: '2',
      firstName: 'Jane2',
      lastName: 'Roy2',
      emailId: 'janeroy@gmail2.com'
      
    },
    
  ];

  constructor(private httpClient: HttpClient) { }

  all(): Observable<Array<object>> {
    return of(this.employee);
  }
  
  
  findOne(id: string): Observable<object> {
    const user = this.users.find((u: any) => {
      return u.id === id;
    });
    return of(user);
  }
  
  getEmployeesList(): Observable<Employee[]>{
    return this.httpClient.get<Employee[]>(`${this.baseURL}`);
  }

  createEmployee(employee: Employee): Observable<Object>{
    return this.httpClient.post(`${this.baseURL}`, employee);
  }

  getEmployeeById(id: number): Observable<Employee>{
    return this.httpClient.get<Employee>(`${this.baseURL}/${id}`);
  }

  updateEmployee(id: number, employee: Employee): Observable<Object>{
    return this.httpClient.put(`${this.baseURL}/${id}`, employee);
  }

  deleteEmployee(id: number): Observable<Object>{
    return this.httpClient.delete(`${this.baseURL}/${id}`);
  }
}
